move to metrics mixin
